Adjusted command update workflow
